Experiment with AdvancedDLSupport
Experiment with using AdvancedDLSupport for P/Invoke implementation.
