+++ src/SqlLocalDb/MartinCostello.SqlLocalDb.csproj
@@ -20,6 +20,7 @@
     <EmbeddedResource Update="SR.resx" Generator="ResXFileCodeGenerator" LastGenOutput="SR.Designer.cs" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="AdvancedDLSupport" Version="2.3.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.0.0" />
     <PackageReference Include="Microsoft.Win32.Registry" Version="4.5.0" />
